git-svn-id: svn://192.168.1.10/public/CoffeeMud@4977 0d6f1817-ed0e-0410-87c9-987e46238f29

+++ com/planet_ink/coffee_mud/common/DefaultPlayerStats.java
@@ -343,7 +343,7 @@ public class DefaultPlayerStats implements PlayerStats
         for(int ss=0;ss<levelDateTimes.size();ss++)
         {
             buf.append(((Integer)levelDateTimes.elementAt(ss,1)).intValue()+",");
-            buf.append(((Long)levelDateTimes.elementAt(ss,2)).intValue()+";");
+            buf.append(((Long)levelDateTimes.elementAt(ss,2)).longValue()+";");
         }
         return buf.toString();
     }
